refactor(twitter): modularize scraper components and improve error handling
- Move account management logic from auth.go to a separate file
- Centralize authentication and rate limit handling in common functions
- Simplify ScrapeFollowersForProfile and ScrapeTweetsByQuery using Retry
- Remove duplicate code and unnecessary initializations
- Increase WorkerResponseTimeout from 30 to 45 seconds in DefaultConfig
