+++ pkg/scrapers/twitter/auth.go
@@ -2,59 +2,12 @@ package twitter
 
 import (
 	"fmt"
-	"sync"
-	"time"
 
-	twitterscraper "github.com/masa-finance/masa-twitter-scraper"
 	"github.com/sirupsen/logrus"
 )
 
-type Scraper struct {
-	*twitterscraper.Scraper
-}
-
-type TwitterAccount struct {
-	Username         string
-	Password         string
-	TwoFACode        string
-	RateLimitedUntil time.Time
-}
-
-type TwitterAccountManager struct {
-	accounts []*TwitterAccount
-	index    int
-	mutex    sync.Mutex
-}
-
-func NewTwitterAccountManager(accounts []*TwitterAccount) *TwitterAccountManager {
-	return &TwitterAccountManager{
-		accounts: accounts,
-		index:    0,
-	}
-}
-
-func (manager *TwitterAccountManager) GetNextAccount() *TwitterAccount {
-	manager.mutex.Lock()
-	defer manager.mutex.Unlock()
-	for i := 0; i < len(manager.accounts); i++ {
-		account := manager.accounts[manager.index]
-		manager.index = (manager.index + 1) % len(manager.accounts)
-		if time.Now().After(account.RateLimitedUntil) {
-			return account
-		}
-	}
-	return nil
-}
-
-func (manager *TwitterAccountManager) MarkAccountRateLimited(account *TwitterAccount) {
-	manager.mutex.Lock()
-	defer manager.mutex.Unlock()
-	account.RateLimitedUntil = time.Now().Add(GetRateLimitDuration())
-}
-
 func NewScraper(account *TwitterAccount, cookieDir string) *Scraper {
-	ts := twitterscraper.New()
-	scraper := &Scraper{Scraper: ts}
+	scraper := &Scraper{Scraper: newTwitterScraper()}
 
 	if err := LoadCookies(scraper.Scraper, account, cookieDir); err == nil {
 		logrus.Debugf("Cookies loaded for user %s.", account.Username)
@@ -99,8 +52,4 @@ func (scraper *Scraper) Logout() error {
 		return fmt.Errorf("logout failed: %v", err)
 	}
 	return nil
-}
-
-func (scraper *Scraper) IsLoggedIn() bool {
-	return scraper.Scraper.IsLoggedIn()
 }
